requirements.txt
fastapi
uvicorn
python-dotenv
requests
websockets
pydub
uuid
venv
 ag2==0.7.2
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4